requirements.txt
tensorflow>=2.12.0
datasets>=2.14.0
transformers>=4.35.0
huggingface-hub>=0.19.0
matplotlib>=3.7.1
pillow>=10.0.0
numpy>=1.23.0
pyyaml>=6.0.1
tqdm>=4.65.0
